keenmind/object-detection/docker-requirements.txt
torch
torchvision
imgaug
fastapi
uvicorn
PyYAML==5.4.1
python-multipart
